requirements.txt
Django==6.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.4.0
python-decouple==3.8
